% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_go.R
\name{get_genes_for_go}
\alias{get_genes_for_go}
\title{Title}
\usage{
get_genes_for_go(go_id, organism)
}
\arguments{
\item{go_id}{Gene Ontology Accession number, such as "GO:0001773"}

\item{organism}{Species : either mouse or human}
}
\value{
genes a vector of genes in this gene ontology for this organism
}
\description{
Title
}
