% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_def_df_with_genes}
\alias{go_def_df_with_genes}
\title{Gene ontology definition data frame with descendant genes belonging to this gene ontology}
\format{
## `go_def.dataframe`
A data frame with  and 6 columns:
\describe{
  \item{id}{gene ontology id, such as  GO:0031330}
  \item{name}{name of gene ontology. Example: negative regulation of cellular catabolic process biological_process}
  \item{namespace}{The category of the gene ontolgy. For example: biological_process.}
  \item{def}{The definition  of the gene ontolgy. For example: The process in which an antigen-presenting cell expresses antigen (peptide or lipid) on its cell surface in association with an MHC protein complex.}
  \item{human}{The human genes that belong to this gene ontology and all sub ontologies.}
  \item{mouse}{The mouse genes that belong to this gene ontology and all sub ontologies.}
}
}
\usage{
go_def_df_with_genes
}
\description{
Gene ontology definition data frame with descendant genes belonging to this gene ontology
}
\keyword{datasets}
